+++ helix/constants.go
@@ -1,5 +1,6 @@
-package conduit
+package helix
 
+// TransportMethod is the method of transport used by the conduit/subscription
 type TransportMethod string
 
 const (
@@ -8,20 +9,13 @@ const (
 	TransportMethodConduit   TransportMethod = "conduit"
 )
 
+// EventType is the type of event that the conduit/subscription is for
 type EventType string
 
 const (
-	EventTypeNull           EventType = ""
-	EventTypeUnknown        EventType = "unknown"
+	// EventTypeNull is an empty event type, and therefore the default value of EventType
+	EventTypeNull EventType = ""
+	//EventTypeChannelMessage is a chat message
 	EventTypeChannelMessage EventType = "channel.chat.message"
 	// TODO: Add more event types
 )
-
-func ParseEventType(s string) EventType {
-	switch s {
-	case "channel.chat.message":
-		return EventTypeChannelMessage
-	default:
-		return EventTypeUnknown
-	}
-}
