+++ helix/helix.go
@@ -1,4 +1,4 @@
-package conduit
+package helix
 
 import (
 	"io"
