+++ helix/parse.go
@@ -1,9 +1,10 @@
-package conduit
+package helix
 
 import (
 	"encoding/json"
 )
 
+// ParseChannelMessage parses a channel message from a slice of bytes
 func ParseChannelMessage(data []byte) (ChannelMessage, error) {
 	var msg ChannelMessage
 	err := json.Unmarshal(data, &msg)
