+++ helix/transport.go
@@ -1,4 +1,4 @@
-package conduit
+package helix
 
 // Transport is used to manage websocket & webhook conduits as a single interface
 type Transport interface {
@@ -6,7 +6,8 @@ type Transport interface {
 	Init() error
 	// Close closes the transport
 	Close()
-	// Ready return the ready channel
+	// Ready returns the ready channel, which is closed when the transport is ready to be used.
+	// Reading from the channel, after calling Init, will block until the transport is ready.
 	Ready() chan struct{}
 	// GetTransportUpdate returns a transport update used to update the conduit
 	GetTransportUpdate() *TransportUpdate
